Add RelativeUrlSettingInput
